Edited pverify*_web.html pages to fix known issue in 5.3 (RT 150113)
Known Issue with 5.3:

DESCRIPTION:
Forms on pverify*_web.html pages take patron to the wrong page after submitting, when using <!--{formbegin}-->

On pverify*_web.html pages, the <!--{formbegin}--> token generates generates HTML that takes patrons to the wrong page when they submit a form. For example, if patrons are prompted to log in while requesting, they could be taken to /search instead of to the next page in the requesting process.

The workaround is to replace the <!--{formbegin}--> token on any pverify*_web.html page that is using it with the following line of code: <form name="patform" target="_parent" method="POST">.

RESOLUTION
Fixed in Sierra 5.4